requirements.txt
# --- Основа бота ---
beautifulsoup4
lxml
aiogram>=3.13.0
aiohttp>=3.9.0
aiofiles>=23.2.1
python-dotenv>=1.0.0
loguru>=0.7.2
notifiers>=1.3.3

# --- База данных ---
SQLAlchemy>=2.0.23
asyncpg>=0.29.0
alembic>=1.13.1
redis>=5.0.1

# --- Работа с расписанием и файлами ---
pdfservices-sdk
pandas
openpyxl
pdfplumber
apscheduler
pytz

# --- Утилиты ---
requests>=2.31.0
aiogram-calendar>=0.1.1
